refacto(server): hydrate reseaux
