+++ .infra/ansible/roles/setup/files/app/tools/cron-data-jobs/run-daily-jobs.sh
@@ -14,8 +14,8 @@ call_daily_jobs_with_logs(){
   # Remplissage des organismes depuis le référentiel
   docker exec flux_retour_cfas_server bash -c "yarn cli hydrate:organismes" || true
   
-  # Remplissage des réseaux depuis csv fournis
-  docker exec flux_retour_cfas_server bash -c "yarn cli hydrate:reseaux-newFormat" || true
+  # Remplissage des réseaux
+  docker exec flux_retour_cfas_server bash -c "yarn cli hydrate:reseaux" || true
 
   # Purge des données inutiles
   docker exec flux_retour_cfas_server bash -c "yarn cli purge:events" || true
